refactor(check): colocate non-lint tests into parse.rs
Move the non-lint test clusters out of the monolithic
crates/qedgen/src/check/tests.rs and into check/parse.rs, next to the
code they exercise (parse_spec_file / parse_spec_file_with_lock and the
import resolver / qed.lock wiring). Move-only: test bodies are
byte-identical modulo indentation; no assertions added or changed.

Clusters moved to check/parse.rs (new `#[cfg(test)] mod tests`):
- import resolution + interface merge (4 tests)
- qed.lock integration (7 tests + the write_simple_path_dep_setup
  helper, used only by moved tests)
- verified_proof_pkgs population (2 tests, share the same helper)

Left in tests.rs: the collect_require_verified_findings tests stay,
because their defining fn lives in check/lints.rs, which is out of
scope for this refactor. All lint-rule tests are untouched.

Test count is preserved (141 total): tests.rs 141 -> 128, parse.rs
0 -> 13. Gates pass: build, cargo test, fmt --check, clippy -D warnings.
