+++ crates/qedgen/src/check/tests.rs
@@ -2805,468 +2805,6 @@ handler deposit (n : U64) : State.Active -> State.Active {
     );
 }
 
-// ----- import resolution + interface merge -----
-
-#[test]
-fn parse_spec_file_resolves_path_imports_and_merges_interface() {
-    let tmp = tempfile::tempdir().unwrap();
-    let spec_dir = tmp.path();
-
-    // Imported interface lives at <dir>/token.qedspec.
-    std::fs::write(
-        spec_dir.join("token.qedspec"),
-        r#"spec SplTokenInterface
-interface Token {
-  program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
-  handler transfer (amount : U64) {
-    discriminant "0x03"
-    accounts {
-      from      : writable, type token
-      to        : writable, type token
-      authority : signer
-    }
-    requires amount > 0
-    ensures  amount > 0
-  }
-}
-"#,
-    )
-    .unwrap();
-
-    // Manifest declares a path source.
-    std::fs::write(
-        spec_dir.join("qed.toml"),
-        r#"
-[dependencies]
-spl_token = { path = "token.qedspec" }
-"#,
-    )
-    .unwrap();
-
-    // Consumer spec imports the interface.
-    let consumer_path = spec_dir.join("escrow.qedspec");
-    std::fs::write(
-        &consumer_path,
-        r#"spec Escrow
-import Token from "spl_token"
-
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-
-    let parsed = parse_spec_file(&consumer_path).expect("parse + resolve should succeed");
-    assert_eq!(parsed.imports.len(), 1);
-    assert_eq!(parsed.imports[0].name, "Token");
-    // Token interface from token.qedspec should now be in parsed.interfaces.
-    assert!(
-        parsed.interfaces.iter().any(|i| i.name == "Token"),
-        "Token interface should be merged into parsed.interfaces; got {:?}",
-        parsed
-            .interfaces
-            .iter()
-            .map(|i| &i.name)
-            .collect::<Vec<_>>(),
-    );
-}
-
-#[test]
-fn parse_spec_file_errors_when_imports_present_but_no_qed_toml() {
-    let tmp = tempfile::tempdir().unwrap();
-    let consumer_path = tmp.path().join("escrow.qedspec");
-    std::fs::write(
-        &consumer_path,
-        r#"spec Escrow
-import Token from "spl_token"
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-
-    let err = format!("{:#}", parse_spec_file(&consumer_path).unwrap_err());
-    assert!(
-        err.contains("no `qed.toml`"),
-        "expected `no qed.toml` error, got: {err}"
-    );
-}
-
-#[test]
-fn parse_spec_file_errors_when_bound_name_not_in_imported_source() {
-    let tmp = tempfile::tempdir().unwrap();
-    let spec_dir = tmp.path();
-
-    std::fs::write(
-        spec_dir.join("other.qedspec"),
-        r#"spec OtherIface
-interface NotToken {
-  program_id "11111111111111111111111111111111"
-}
-"#,
-    )
-    .unwrap();
-    std::fs::write(
-        spec_dir.join("qed.toml"),
-        r#"
-[dependencies]
-spl_token = { path = "other.qedspec" }
-"#,
-    )
-    .unwrap();
-    let consumer_path = spec_dir.join("escrow.qedspec");
-    std::fs::write(
-        &consumer_path,
-        r#"spec Escrow
-import Token from "spl_token"
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-
-    let err = format!("{:#}", parse_spec_file(&consumer_path).unwrap_err());
-    assert!(
-        err.contains("declares no `interface Token`"),
-        "expected `no interface Token` error, got: {err}"
-    );
-}
-
-#[test]
-fn parse_spec_file_no_imports_does_not_require_qed_toml() {
-    let tmp = tempfile::tempdir().unwrap();
-    let path = tmp.path().join("plain.qedspec");
-    std::fs::write(
-        &path,
-        r#"spec Plain
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-    // No qed.toml, no imports — should parse cleanly.
-    let parsed = parse_spec_file(&path).unwrap();
-    assert!(parsed.imports.is_empty());
-}
-
-// ----- qed.lock integration -----
-
-fn write_simple_path_dep_setup(spec_dir: &std::path::Path) -> std::path::PathBuf {
-    std::fs::write(
-        spec_dir.join("token.qedspec"),
-        r#"spec SplTokenInterface
-interface Token {
-  program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
-  upstream {
-    package      "spl-token"
-    version      "4.0.3"
-    binary_hash  "sha256:9c1edeadbeef"
-    verified_with ["proptest"]
-    verified_at  "2026-04-25"
-  }
-  handler transfer (amount : U64) {
-    discriminant "0x03"
-    accounts {
-      from      : writable, type token
-      to        : writable, type token
-      authority : signer
-    }
-    requires amount > 0
-    ensures  amount > 0
-  }
-}
-"#,
-    )
-    .unwrap();
-    std::fs::write(
-        spec_dir.join("qed.toml"),
-        r#"
-[dependencies]
-spl_token = { path = "token.qedspec" }
-"#,
-    )
-    .unwrap();
-    let consumer = spec_dir.join("escrow.qedspec");
-    std::fs::write(
-        &consumer,
-        r#"spec Escrow
-import Token from "spl_token"
-
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-    consumer
-}
-
-#[test]
-fn parse_spec_file_auto_writes_lock_with_resolved_entries() {
-    let tmp = tempfile::tempdir().unwrap();
-    let consumer = write_simple_path_dep_setup(tmp.path());
-
-    // Lock should not exist before parse.
-    assert!(!tmp.path().join("qed.lock").exists());
-
-    parse_spec_file(&consumer).expect("parse should succeed and write lock");
-
-    let lock = crate::qed_lock::read(tmp.path())
-        .unwrap()
-        .expect("lock should be written");
-    assert_eq!(lock.dependencies.len(), 1);
-    let entry = &lock.dependencies[0];
-    assert_eq!(entry.name, "spl_token");
-    assert_eq!(entry.source, "path:token.qedspec");
-    assert!(entry.spec_hash.starts_with("sha256:"));
-    // Path source — no commit, no ref, no sub-path.
-    assert!(entry.git_ref.is_none());
-    assert!(entry.resolved_commit.is_none());
-    // Upstream block from the imported interface flowed through.
-    assert_eq!(
-        entry.upstream_binary_hash.as_deref(),
-        Some("sha256:9c1edeadbeef")
-    );
-    assert_eq!(entry.upstream_version.as_deref(), Some("4.0.3"));
-}
-
-#[test]
-fn parse_spec_file_with_lock_frozen_errors_when_lock_missing() {
-    let tmp = tempfile::tempdir().unwrap();
-    let consumer = write_simple_path_dep_setup(tmp.path());
-
-    // Frozen mode + no lock on disk → error.
-    let err = format!(
-        "{:#}",
-        parse_spec_file_with_lock(&consumer, crate::qed_lock::LockMode::Frozen).unwrap_err()
-    );
-    assert!(err.contains("stale (--frozen)"), "got: {err}");
-}
-
-#[test]
-fn parse_spec_file_with_lock_frozen_succeeds_when_lock_current() {
-    let tmp = tempfile::tempdir().unwrap();
-    let consumer = write_simple_path_dep_setup(tmp.path());
-
-    // Auto first to write the lock, then Frozen to verify it stays current.
-    parse_spec_file(&consumer).unwrap();
-    parse_spec_file_with_lock(&consumer, crate::qed_lock::LockMode::Frozen)
-        .expect("frozen should pass when lock is current");
-}
-
-#[test]
-fn parse_spec_file_renames_imported_interface_via_as_alias() {
-    let tmp = tempfile::tempdir().unwrap();
-    let spec_dir = tmp.path();
-
-    std::fs::write(
-        spec_dir.join("token.qedspec"),
-        r#"spec SplTokenInterface
-interface Token {
-  program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
-  handler transfer (amount : U64) {
-    discriminant "0x03"
-    accounts {
-      from      : writable, type token
-      to        : writable, type token
-      authority : signer
-    }
-    requires amount > 0
-    ensures  amount > 0
-  }
-}
-"#,
-    )
-    .unwrap();
-    std::fs::write(
-        spec_dir.join("qed.toml"),
-        r#"
-[dependencies]
-spl_token = { path = "token.qedspec" }
-"#,
-    )
-    .unwrap();
-    // Consumer uses `as Tk` to rename Token → Tk in its namespace.
-    let consumer = spec_dir.join("escrow.qedspec");
-    std::fs::write(
-        &consumer,
-        r#"spec Escrow
-import Token from "spl_token" as Tk
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-
-    let parsed = parse_spec_file(&consumer).expect("alias-renamed import should parse + merge");
-    // Imported interface should appear under its alias name `Tk`,
-    // not the source-side `Token`.
-    assert!(
-        parsed.interfaces.iter().any(|i| i.name == "Tk"),
-        "expected interface renamed to `Tk`; got {:?}",
-        parsed
-            .interfaces
-            .iter()
-            .map(|i| &i.name)
-            .collect::<Vec<_>>(),
-    );
-    assert!(
-        !parsed.interfaces.iter().any(|i| i.name == "Token"),
-        "the source-side name `Token` should not leak into consumer when an alias is set"
-    );
-}
-
-#[test]
-fn parse_spec_file_resolves_multi_file_imported_dep() {
-    let tmp = tempfile::tempdir().unwrap();
-    let spec_dir = tmp.path();
-
-    // Imported dep is a *directory* of fragments. Each declares the
-    // same `spec MultiToken`; one carries the interface, another
-    // carries a sidecar event used in the interface's docs.
-    let dep_dir = spec_dir.join("multitoken");
-    std::fs::create_dir(&dep_dir).unwrap();
-    std::fs::write(
-        dep_dir.join("a-iface.qedspec"),
-        r#"spec MultiToken
-interface Token {
-  program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
-  handler transfer (amount : U64) {
-    discriminant "0x03"
-    accounts {
-      from      : writable, type token
-      to        : writable, type token
-      authority : signer
-    }
-    requires amount > 0
-    ensures  amount > 0
-  }
-}
-"#,
-    )
-    .unwrap();
-    std::fs::write(
-        dep_dir.join("b-event.qedspec"),
-        r#"spec MultiToken
-event TokenMoved {
-  amount : U64,
-}
-"#,
-    )
-    .unwrap();
-
-    std::fs::write(
-        spec_dir.join("qed.toml"),
-        r#"
-[dependencies]
-spl_token = { path = "multitoken" }
-"#,
-    )
-    .unwrap();
-
-    let consumer = spec_dir.join("escrow.qedspec");
-    std::fs::write(
-        &consumer,
-        r#"spec Escrow
-import Token from "spl_token"
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-
-    let parsed = parse_spec_file(&consumer)
-        .expect("multi-file imported dep should parse + merge end-to-end");
-    // Token interface from a-iface.qedspec lives in the merged consumer.
-    assert!(
-        parsed.interfaces.iter().any(|i| i.name == "Token"),
-        "interface from multi-file dep should be merged in; got {:?}",
-        parsed
-            .interfaces
-            .iter()
-            .map(|i| &i.name)
-            .collect::<Vec<_>>(),
-    );
-}
-
-#[test]
-fn parse_spec_file_errors_when_multi_file_dep_fragments_disagree_on_spec_name() {
-    let tmp = tempfile::tempdir().unwrap();
-    let spec_dir = tmp.path();
-
-    let dep_dir = spec_dir.join("bad-multi");
-    std::fs::create_dir(&dep_dir).unwrap();
-    std::fs::write(
-        dep_dir.join("a.qedspec"),
-        "spec NameOne\ninterface Token { program_id \"x\" }\n",
-    )
-    .unwrap();
-    std::fs::write(
-        dep_dir.join("b.qedspec"),
-        "spec NameTwo\nevent E { amount : U64 }\n",
-    )
-    .unwrap();
-
-    std::fs::write(
-        spec_dir.join("qed.toml"),
-        r#"
-[dependencies]
-bad = { path = "bad-multi" }
-"#,
-    )
-    .unwrap();
-
-    let consumer = spec_dir.join("c.qedspec");
-    std::fs::write(
-        &consumer,
-        r#"spec Caller
-import Token from "bad"
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-
-    let err = format!("{:#}", parse_spec_file(&consumer).unwrap_err());
-    assert!(
-        err.contains("must declare the same name"),
-        "expected name-mismatch error; got: {err}"
-    );
-}
-
-#[test]
-fn parse_spec_file_with_lock_frozen_errors_when_imported_source_changed() {
-    let tmp = tempfile::tempdir().unwrap();
-    let consumer = write_simple_path_dep_setup(tmp.path());
-
-    // Auto-write a baseline lock, then mutate the imported source — the
-    // spec hash should drift, so Frozen catches it.
-    parse_spec_file(&consumer).unwrap();
-    std::fs::write(
-        tmp.path().join("token.qedspec"),
-        r#"spec SplTokenInterface
-interface Token {
-  program_id "DIFFERENT11111111111111111111111111111111"
-  handler transfer (amount : U64) {
-    discriminant "0x03"
-    accounts {
-      from      : writable, type token
-      to        : writable, type token
-      authority : signer
-    }
-    requires amount > 0
-    ensures  amount > 0
-  }
-}
-"#,
-    )
-    .unwrap();
-    let err = format!(
-        "{:#}",
-        parse_spec_file_with_lock(&consumer, crate::qed_lock::LockMode::Frozen).unwrap_err()
-    );
-    assert!(err.contains("spec_hash"), "got: {err}");
-}
-
 // ----- Rule 17: invariant_no_body -----
 
 #[test]
@@ -4545,90 +4083,3 @@ handler pay : State.A -> State.A {
 }
 
 // ----- end collect_require_verified_findings -----
-
-// ------------------------------------------------------------------
-// ParsedSpec.verified_proof_pkgs population — pins the
-// resolver→ParsedSpec wiring (the `lake build` runner is exercised
-// end-to-end elsewhere).
-// ------------------------------------------------------------------
-
-#[test]
-fn verified_proof_pkgs_populated_when_provider_ships_proof_package() {
-    let tmp = tempfile::tempdir().unwrap();
-    let spec_dir = tmp.path();
-
-    // Provider qedspec at spec_dir/token.qedspec.
-    std::fs::write(
-        spec_dir.join("token.qedspec"),
-        r#"spec TokenLib
-interface Token {
-  program_id "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
-  handler transfer (amount : U64) {
-    accounts {
-      from      : writable
-      to        : writable
-      authority : signer
-    }
-    ensures amount > 0
-  }
-}
-"#,
-    )
-    .unwrap();
-    // Proof package alongside the qedspec — both module + lakefile
-    // must be present for `has_proofs` to be true.
-    let proofs_dir = spec_dir.join(".qed").join("proofs");
-    std::fs::create_dir_all(&proofs_dir).unwrap();
-    std::fs::write(proofs_dir.join("Token.lean"), "-- stub proof").unwrap();
-    std::fs::write(proofs_dir.join("lakefile.lean"), "package tokenProofs").unwrap();
-
-    std::fs::write(
-        spec_dir.join("qed.toml"),
-        r#"
-[dependencies]
-spl_token = { path = "token.qedspec" }
-"#,
-    )
-    .unwrap();
-    let consumer = spec_dir.join("escrow.qedspec");
-    std::fs::write(
-        &consumer,
-        r#"spec Escrow
-import Token from "spl_token"
-
-type State | A of { x : U64 }
-handler h : State.A -> State.A { effect { x := 1 } }
-"#,
-    )
-    .unwrap();
-
-    let parsed = parse_spec_file(&consumer).expect("parse should succeed");
-    assert_eq!(
-        parsed.verified_proof_pkgs.len(),
-        1,
-        "expected 1 proof package; got {:?}",
-        parsed.verified_proof_pkgs
-    );
-    assert!(
-        parsed.verified_proof_pkgs[0].ends_with(".qed/proofs")
-            || parsed.verified_proof_pkgs[0].ends_with(".qed\\proofs"),
-        "should point at the provider's proof package root; got: {}",
-        parsed.verified_proof_pkgs[0].display()
-    );
-}
-
-#[test]
-fn verified_proof_pkgs_empty_when_no_provider_proofs() {
-    // No `.qed/proofs/` alongside the provider qedspec → resolver
-    // sets has_proofs=false → no entry in verified_proof_pkgs.
-    let tmp = tempfile::tempdir().unwrap();
-    let consumer = write_simple_path_dep_setup(tmp.path());
-    let parsed = parse_spec_file(&consumer).expect("parse should succeed");
-    assert!(
-        parsed.verified_proof_pkgs.is_empty(),
-        "no provider proofs → empty list; got: {:?}",
-        parsed.verified_proof_pkgs
-    );
-}
-
-// ----- end verified_proof_pkgs -----
